+++ inc/lilyNotation/index.ts
@@ -2,7 +2,12 @@
 import {Note, Notation, assignNotationEventsIds} from "./notation";
 import {fuzzyMatchNotations} from "./fuzzyMatch";
 import {matchWithMIDI, matchWithExactMIDI} from "./matcher";
+import * as measureLayout from "./measureLayout";
 import {LayoutType} from "./measureLayout";
+
+
+
+const {...MLayoutClasses} = measureLayout;
 
 
 
@@ -14,4 +19,5 @@ export {
 	matchWithMIDI,
 	matchWithExactMIDI,
 	LayoutType,
+	MLayoutClasses,
 };